HPCC-20525 Avoid static library build issues
Instead of using a static library for cryptohelpe, include the
sources in jlib's cmake file.

+++ system/security/cryptohelper/cryptocommon.cpp
@@ -32,7 +32,7 @@ static std::vector<std::unique_ptr<CriticalSection>> cryptoLocks;
 
 static void locking_function(int mode, int n, const char * file, int line)
 {
-    assertex(n < numCryptoLocks);
+    assertex(n < (int)numCryptoLocks);
     if (mode & CRYPTO_LOCK)
         cryptoLocks[n]->enter();
     else
